1) changes in AccountController: - add post route - small fixes in delete route 2) delete fields createdDate and AccountHolderName from Account 3) add dialect in configuration in GuiceModule 4) add foreign key "userId" in Account

+++ src/main/java/orr/controller/AccountController.java
@@ -3,11 +3,18 @@ package orr.controller;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.inject.Inject;
+import orr.adapter.LocalDateTypeAdapter;
 import orr.dto.AccountDto;
 import orr.errors.ResponseError;
 import orr.exception.AccountNotFoundException;
 import orr.models.Account;
+import orr.models.User;
 import orr.service.Impl.AccountServiceImpl;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Locale;
 
 import static orr.utils.JsonUtil.json;
 import static orr.utils.JsonUtil.toJson;
@@ -21,35 +28,58 @@ public class AccountController {
 
         get("/accounts/:id", (req, res) -> {
             Long id = Long.valueOf(req.params(":id"));
-            Account account = accountService.getById(id);
-            if (account != null) {
-                return account;
+
+            if (!accountService.findById(id).isPresent()) {
+                res.status(400);
+                return new ResponseError("No account with id '%s' found", String.valueOf(id));
             }
-            res.status(400);
-            return new ResponseError("No account with id '%s' found", String.valueOf(id));
+
+            res.status(200);
+            return accountService.getById(id);
+        }, json());
+
+        post("/accounts/:userId", (req, res) -> {
+            Long userId = Long.valueOf(req.params(":userId"));
+
+            String request = "" + req.body();
+            Gson gson = new GsonBuilder().create();
+            Account account = gson.fromJson(request, Account.class);
+
+            Account newAccount = accountService.add(userId, account);
+            if (account != null) {
+                return newAccount;
+            }
+            return new ResponseError("No user created");
         }, json());
 
         put("/accounts/:id", (req, res) -> {
             Long id = Long.valueOf(req.params(":id"));
+            if (!accountService.findById(id).isPresent()) {
+                res.status(400);
+                return new ResponseError("No account with id '%s' found", String.valueOf(id));
+            }
             String request = "" + req.body();
             Gson gson = new GsonBuilder().create();
             AccountDto accountDto = gson.fromJson(request, AccountDto.class);
-            Account account = accountService.update(id, accountDto);
-            if (account != null) {
-                return account;
-            }
-            return new ResponseError("No account with id '%s' found", String.valueOf(id));
+            res.status(200);
+            return accountService.update(id, accountDto);
         }, json());
 
         delete("/accounts/:id", (req, res) -> {
             Long id = Long.valueOf(req.params(":id"));
-            try {
-                accountService.delete(id);
+            if (!accountService.findById(id).isPresent()) {
                 res.status(400);
-            } catch (Exception e){
-                throw new AccountNotFoundException();
+                return new ResponseError("No account with id '%s' found", String.valueOf(id));
             }
-            return new ResponseError("No account with id '%s' found", String.valueOf(id));
+            accountService.delete(id);
+            res.status(200);
+            return "Account was deleted.";
+        }, json());
+
+        get("/accounts/user/:id", (req, res) -> {
+            Long userId = Long.valueOf(":id");
+            res.status(200);
+            return accountService.getByUserId(userId);
         }, json());
 
         after((req, res) -> {
